+++ rewatch/src/build.rs
@@ -18,10 +18,12 @@ use console::style;
 use indicatif::{ProgressBar, ProgressStyle};
 use log::log_enabled;
 use serde::Serialize;
+use std::ffi::OsString;
 use std::fmt;
 use std::fs::File;
 use std::io::{stdout, Write};
 use std::path::{Path, PathBuf};
+use std::process::Stdio;
 use std::time::{Duration, Instant};
 
 use self::compile::compiler_args;
@@ -57,7 +59,7 @@ pub struct CompilerArgs {
 pub fn get_compiler_args(
     path: &Path,
     rescript_version: Option<String>,
-    bsc_path: &Option<PathBuf>,
+    bsc_path: Option<PathBuf>,
     build_dev_deps: bool,
 ) -> Result<String> {
     let filename = &helpers::get_abs_path(path);
@@ -499,7 +501,7 @@ pub fn build(
     show_progress: bool,
     no_timing: bool,
     create_sourcedirs: bool,
-    bsc_path: &Option<PathBuf>,
+    bsc_path: Option<PathBuf>,
     build_dev_deps: bool,
     snapshot_output: bool,
 ) -> Result<BuildState> {
@@ -514,7 +516,7 @@ pub fn build(
         filter,
         show_progress,
         path,
-        bsc_path,
+        &bsc_path,
         build_dev_deps,
         snapshot_output,
     )
@@ -548,6 +550,28 @@ pub fn build(
             clean::cleanup_after_build(&build_state);
             write_build_ninja(&build_state);
             Err(anyhow!("Incremental build failed. Error: {e}"))
+        }
+    }
+}
+
+pub fn pass_through_legacy(mut args: Vec<OsString>) -> i32 {
+    let project_root = helpers::get_abs_path(Path::new("."));
+    let workspace_root = helpers::get_workspace_root(&project_root);
+
+    let rescript_legacy_path = helpers::get_rescript_legacy(&project_root, workspace_root);
+
+    args.insert(0, rescript_legacy_path.into());
+    let status = std::process::Command::new("node")
+        .args(args)
+        .stdout(Stdio::inherit())
+        .stderr(Stdio::inherit())
+        .status();
+
+    match status {
+        Ok(s) => s.code().unwrap_or(0),
+        Err(err) => {
+            eprintln!("Error running the legacy build system: {err}");
+            1
         }
     }
 }