+++ rewatch/src/helpers.rs
@@ -181,32 +181,35 @@ pub fn create_path_for_path(path: &Path) {
     fs::DirBuilder::new().recursive(true).create(path).unwrap();
 }
 
-pub fn get_bsc(root_path: &Path, workspace_root: &Option<PathBuf>) -> PathBuf {
+fn get_bin_dir() -> PathBuf {
     let subfolder = match (std::env::consts::OS, std::env::consts::ARCH) {
         ("macos", "aarch64") => "darwin-arm64",
         ("macos", _) => "darwin-x64",
         ("linux", "aarch64") => "linux-arm64",
         ("linux", _) => "linux-x64",
         ("windows", "aarch64") => "win-arm64",
-        ("windows", _) => "win-x64",
+        ("windows", _) => "win32-x64",
         _ => panic!("Unsupported architecture"),
     };
 
+    Path::new("node_modules")
+        .join("@rescript")
+        .join(subfolder)
+        .join("bin")
+}
+
+pub fn get_bsc(root_path: &Path, workspace_root: &Option<PathBuf>) -> PathBuf {
+    let bin_dir = get_bin_dir();
+
     match (
         root_path
-            .join("node_modules")
-            .join("@rescript")
-            .join(subfolder)
-            .join("bin")
+            .join(&bin_dir)
             .join("bsc.exe")
             .canonicalize()
             .map(StrippedVerbatimPath::to_stripped_verbatim_path),
         workspace_root.as_ref().map(|workspace_root| {
             workspace_root
-                .join("node_modules")
-                .join("@rescript")
-                .join(subfolder)
-                .join("bin")
+                .join(&bin_dir)
                 .join("bsc.exe")
                 .canonicalize()
                 .map(StrippedVerbatimPath::to_stripped_verbatim_path)
@@ -215,6 +218,29 @@ pub fn get_bsc(root_path: &Path, workspace_root: &Option<PathBuf>) -> PathBuf {
         (Ok(path), _) => path,
         (_, Some(Ok(path))) => path,
         _ => panic!("Could not find bsc.exe"),
+    }
+}
+
+pub fn get_rescript_legacy(root_path: &Path, workspace_root: Option<PathBuf>) -> PathBuf {
+    let bin_dir = Path::new("node_modules").join("rescript").join("cli");
+
+    match (
+        root_path
+            .join(&bin_dir)
+            .join("rescript.js")
+            .canonicalize()
+            .map(StrippedVerbatimPath::to_stripped_verbatim_path),
+        workspace_root.map(|workspace_root| {
+            workspace_root
+                .join(&bin_dir)
+                .join("rescript.js")
+                .canonicalize()
+                .map(StrippedVerbatimPath::to_stripped_verbatim_path)
+        }),
+    ) {
+        (Ok(path), _) => path,
+        (_, Some(Ok(path))) => path,
+        _ => panic!("Could not find rescript.exe"),
     }
 }