+++ rewatch/src/main.rs
@@ -1,93 +1,17 @@
 use anyhow::Result;
-use clap::{Parser, ValueEnum};
-use clap_verbosity_flag::InfoLevel;
+use clap::Parser;
 use log::LevelFilter;
 use regex::Regex;
-use std::io::Write;
-use std::path::{Path, PathBuf};
+use std::{
+    io::Write,
+    path::{Path, PathBuf},
+};
 
-use rewatch::{build, cmd, lock, watcher};
-
-#[derive(Debug, Clone, ValueEnum)]
-enum Command {
-    /// Build using Rewatch
-    Build,
-    /// Build, then start a watcher
-    Watch,
-    /// Clean the build artifacts
-    Clean,
-}
-
-/// Rewatch is an alternative build system for the Rescript Compiler bsb (which uses Ninja internally). It strives
-/// to deliver consistent and faster builds in monorepo setups with multiple packages, where the
-/// default build system fails to pick up changed interfaces across multiple packages.
-#[derive(Parser, Debug)]
-#[command(version)]
-struct Args {
-    #[arg(value_enum)]
-    command: Option<Command>,
-
-    /// The relative path to where the main rescript.json resides. IE - the root of your project.
-    folder: Option<String>,
-
-    /// Filter allows for a regex to be supplied which will filter the files to be compiled. For
-    /// instance, to filter out test files for compilation while doing feature work.
-    #[arg(short, long)]
-    filter: Option<String>,
-
-    /// This allows one to pass an additional command to the watcher, which allows it to run when
-    /// finished. For instance, to play a sound when done compiling, or to run a test suite.
-    /// NOTE - You may need to add '--color=always' to your subcommand in case you want to output
-    /// colour as well
-    #[arg(short, long)]
-    after_build: Option<String>,
-
-    // Disable timing on the output
-    #[arg(short, long, default_value = "false", num_args = 0..=1)]
-    no_timing: bool,
-
-    // simple output for snapshot testing
-    #[arg(short, long, default_value = "false", num_args = 0..=1)]
-    snapshot_output: bool,
-
-    /// Verbosity:
-    /// -v -> Debug
-    /// -vv -> Trace
-    /// -q -> Warn
-    /// -qq -> Error
-    /// -qqq -> Off.
-    /// Default (/ no argument given): 'info'
-    #[command(flatten)]
-    verbose: clap_verbosity_flag::Verbosity<InfoLevel>,
-
-    /// This creates a source_dirs.json file at the root of the monorepo, which is needed when you
-    /// want to use Reanalyze
-    #[arg(short, long, default_value_t = false, num_args = 0..=1)]
-    create_sourcedirs: bool,
-
-    /// This prints the compiler arguments. It expects the path to a rescript.json file.
-    /// This also requires --bsc-path and --rescript-version to be present
-    #[arg(long)]
-    compiler_args: Option<String>,
-
-    /// This is the flag to also compile development dependencies
-    /// It's important to know that we currently do not discern between project src, and
-    /// dependencies. So enabling this flag will enable building _all_ development dependencies of
-    /// _all_ packages
-    #[arg(long, default_value_t = false, num_args = 0..=1)]
-    dev: bool,
-
-    /// To be used in conjunction with compiler_args
-    #[arg(long)]
-    rescript_version: Option<String>,
-
-    /// A custom path to bsc
-    #[arg(long)]
-    bsc_path: Option<String>,
-}
+use rewatch::{build, cli, cmd, lock, watcher};
 
 fn main() -> Result<()> {
-    let args = Args::parse();
+    let args = cli::Cli::parse();
+
     let log_level_filter = args.verbose.log_level_filter();
 
     env_logger::Builder::new()
@@ -96,82 +20,117 @@ fn main() -> Result<()> {
         .target(env_logger::fmt::Target::Stdout)
         .init();
 
-    let command = args.command.unwrap_or(Command::Build);
-    let folder = args.folder.unwrap_or(".".to_string());
-    let filter = args
-        .filter
-        .map(|filter| Regex::new(filter.as_ref()).expect("Could not parse regex"));
-
-    match args.compiler_args {
-        None => (),
-        Some(path) => {
-            println!(
-                "{}",
-                build::get_compiler_args(
-                    Path::new(&path),
-                    args.rescript_version,
-                    &args.bsc_path.map(PathBuf::from),
-                    args.dev
-                )?
-            );
-            std::process::exit(0);
-        }
-    }
+    let command = args.command.unwrap_or(cli::Command::Build(args.build_args));
 
     // The 'normal run' mode will show the 'pretty' formatted progress. But if we turn off the log
     // level, we should never show that.
     let show_progress = log_level_filter == LevelFilter::Info;
 
-    match lock::get(&folder) {
-        lock::Lock::Error(ref e) => {
-            println!("Could not start Rewatch: {e}");
-            std::process::exit(1)
+    match command.clone() {
+        cli::Command::CompilerArgs {
+            path,
+            dev,
+            rescript_version,
+            bsc_path,
+        } => {
+            println!(
+                "{}",
+                build::get_compiler_args(
+                    Path::new(&path),
+                    rescript_version,
+                    bsc_path.as_ref().map(PathBuf::from),
+                    *dev
+                )?
+            );
+            std::process::exit(0);
         }
-        lock::Lock::Aquired(_) => match command {
-            Command::Clean => build::clean::clean(
-                Path::new(&folder),
-                show_progress,
-                &args.bsc_path.map(PathBuf::from),
-                args.dev,
-                args.snapshot_output,
-            ),
-            Command::Build => {
-                match build::build(
-                    &filter,
-                    Path::new(&folder),
-                    show_progress,
-                    args.no_timing,
-                    args.create_sourcedirs,
-                    &args.bsc_path.map(PathBuf::from),
-                    args.dev,
-                    args.snapshot_output,
-                ) {
-                    Err(e) => {
-                        println!("{e}");
-                        std::process::exit(1)
-                    }
-                    Ok(_) => {
-                        if let Some(args_after_build) = args.after_build {
-                            cmd::run(args_after_build)
-                        }
-                        std::process::exit(0)
-                    }
-                };
-            }
-            Command::Watch => {
-                watcher::start(
-                    &filter,
-                    show_progress,
-                    &folder,
-                    args.after_build,
-                    args.create_sourcedirs,
-                    args.dev,
-                    args.bsc_path,
-                    args.snapshot_output,
-                );
+        cli::Command::Build(build_args) => {
+            let _lock = get_lock(&build_args.folder);
 
-                Ok(())
-            }
-        },
+            let filter = build_args
+                .filter
+                .as_ref()
+                .map(|filter| Regex::new(&filter).expect("Could not parse regex"));
+
+            match build::build(
+                &filter,
+                Path::new(&build_args.folder as &str),
+                show_progress,
+                build_args.no_timing,
+                *build_args.create_sourcedirs,
+                build_args.bsc_path.as_ref().map(PathBuf::from),
+                *build_args.dev,
+                *build_args.snapshot_output,
+            ) {
+                Err(e) => {
+                    println!("{e}");
+                    std::process::exit(1)
+                }
+                Ok(_) => {
+                    if let Some(args_after_build) = (*build_args.after_build).clone() {
+                        cmd::run(args_after_build)
+                    }
+                    std::process::exit(0)
+                }
+            };
+        }
+        cli::Command::Watch(watch_args) => {
+            let _lock = get_lock(&watch_args.folder);
+
+            let filter = watch_args
+                .filter
+                .as_ref()
+                .map(|filter| Regex::new(&filter).expect("Could not parse regex"));
+            watcher::start(
+                &filter,
+                show_progress,
+                &watch_args.folder,
+                (*watch_args.after_build).clone(),
+                *watch_args.create_sourcedirs,
+                *watch_args.dev,
+                (*watch_args.bsc_path).clone(),
+                *watch_args.snapshot_output,
+            );
+
+            Ok(())
+        }
+        cli::Command::Clean {
+            folder,
+            bsc_path,
+            snapshot_output,
+        } => {
+            let _lock = get_lock(&folder);
+
+            build::clean::clean(
+                Path::new(&folder as &str),
+                show_progress,
+                bsc_path.as_ref().map(PathBuf::from),
+                *snapshot_output,
+            )
+        }
+        cli::Command::Legacy { legacy_args } => {
+            let code = build::pass_through_legacy(legacy_args);
+            std::process::exit(code);
+        }
+        cli::Command::Format { mut format_args } => {
+            format_args.insert(0, "format".into());
+            let code = build::pass_through_legacy(format_args);
+            std::process::exit(code);
+        }
+        cli::Command::Dump { mut dump_args } => {
+            dump_args.insert(0, "dump".into());
+            let code = build::pass_through_legacy(dump_args);
+            std::process::exit(code);
+        }
+    }
+}
+
+fn get_lock(folder: &str) -> lock::Lock {
+    match lock::get(folder) {
+        lock::Lock::Error(error) => {
+            println!("Could not start Rewatch: {error}");
+            std::process::exit(1);
+        }
+        acquired_lock => acquired_lock,
     }
 }