+++ rewatch/tests/legacy-utils.sh
@@ -1,0 +1,3 @@
+source "utils.sh"
+
+rewatch_legacy() { RUST_BACKTRACE=1 "../../$REWATCH_EXECUTABLE" legacy $@; }