+++ rewatch/tests/suite-ci.sh
@@ -8,10 +8,13 @@ cd $(dirname $0)
 # Get rewatch executable location from the first argument or use default
 if [ -n "$1" ]; then
   REWATCH_EXECUTABLE="$1"
+  BSC_PATH=""
 else
-  REWATCH_EXECUTABLE="../target/release/rewatch --bsc-path ../../_build/install/default/bin/bsc"
+  REWATCH_EXECUTABLE="../target/release/rewatch"
+  BSC_PATH="--bsc-path ../../_build/install/default/bin/bsc"
 fi
 export REWATCH_EXECUTABLE
+export BSC_PATH
 
 source ./utils.sh
 
@@ -37,4 +40,4 @@ else
   exit 1
 fi
 
-./compile.sh && ./watch.sh && ./lock.sh && ./suffix.sh
+./compile.sh && ./watch.sh && ./lock.sh && ./suffix.sh && ./legacy.sh